(Hopefully) Working KB.
Doesn't do anything yet. But it stores knowledge. Basic things like bumps and arrow tracking are in there using prepositional logic. The skeleton for the first-order rule engine is there so we can write rules to infer the locations of other things based on percepts that arent handled by the KB. Everything is slowly starting to come together (hopefully). I neeeeeeeed help.